# backend/internal/presentation/api/server.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// 全ての借り入れの取得
	// (GET /borrowings)
	BorrowingGetAll(ctx echo.Context) error
	// 債権一覧の取得
	// (GET /credits)
	CreditsList(ctx echo.Context) error
	// グループ一覧の取得
	// (GET /groups)
	GroupGetAll(ctx echo.Context) error
	// グループの作成
	// (POST /groups)
	GroupCreate(ctx echo.Context) error
	// グループの取得
	// (GET /groups/{id})
	GroupGet(ctx echo.Context, id string) error
	// グループの更新
	// (PUT /groups/{id})
	GroupUpdate(ctx echo.Context, id string) error
	// グループメンバー一覧の取得
	// (GET /groups/{id}/members)
	GroupGetMembers(ctx echo.Context, id string) error
	// グループメンバーの追加
	// (POST /groups/{id}/members)
	GroupAddMember(ctx echo.Context, id string) error
	// ヘルスチェック
	// (GET /health)
	HealthCheck(ctx echo.Context) error
	// 全ての立て替えの取得
	// (GET /lendings)
	LendingGetAll(ctx echo.Context) error
	// 立て替えの作成
	// (POST /lendings)
	LendingCreate(ctx echo.Context) error
	// 立て替えの削除
	// (DELETE /lendings/{id})
	LendingDelete(ctx echo.Context, id string) error
	// 立て替えの取得
	// (GET /lendings/{id})
	LendingGet(ctx echo.Context, id string) error
	// 立て替えの更新
	// (PUT /lendings/{id})
	LendingUpdate(ctx echo.Context, id string) error
	// 全ての返済の取得
	// (GET /repayments)
	RepaymentGetAll(ctx echo.Context) error
	// 返済の作成
	// (POST /repayments)
	RepaymentCreate(ctx echo.Context) error
	// ユーザー検索
	// (GET /users)
	UserSearch(ctx echo.Context, params UserSearchParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// BorrowingGetAll converts echo context to params.
func (w *ServerInterfaceWrapper) BorrowingGetAll(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.BorrowingGetAll(ctx)
	return err
}

// CreditsList converts echo context to params.
func (w *ServerInterfaceWrapper) CreditsList(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreditsList(ctx)
	return err
}

// GroupGetAll converts echo context to params.
func (w *ServerInterfaceWrapper) GroupGetAll(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GroupGetAll(ctx)
	return err
}

// GroupCreate converts echo context to params.
func (w *ServerInterfaceWrapper) GroupCreate(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GroupCreate(ctx)
	return err
}

// GroupGet converts echo context to params.
func (w *ServerInterfaceWrapper) GroupGet(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GroupGet(ctx, id)
	return err
}

// GroupUpdate converts echo context to params.
func (w *ServerInterfaceWrapper) GroupUpdate(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GroupUpdate(ctx, id)
	return err
}

// GroupGetMembers converts echo context to params.
func (w *ServerInterfaceWrapper) GroupGetMembers(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GroupGetMembers(ctx, id)
	return err
}

// GroupAddMember converts echo context to params.
func (w *ServerInterfaceWrapper) GroupAddMember(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GroupAddMember(ctx, id)
	return err
}

// HealthCheck converts echo context to params.
func (w *ServerInterfaceWrapper) HealthCheck(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.HealthCheck(ctx)
	return err
}

// LendingGetAll converts echo context to params.
func (w *ServerInterfaceWrapper) LendingGetAll(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.LendingGetAll(ctx)
	return err
}

// LendingCreate converts echo context to params.
func (w *ServerInterfaceWrapper) LendingCreate(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.LendingCreate(ctx)
	return err
}

// LendingDelete converts echo context to params.
func (w *ServerInterfaceWrapper) LendingDelete(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.LendingDelete(ctx, id)
	return err
}

// LendingGet converts echo context to params.
func (w *ServerInterfaceWrapper) LendingGet(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.LendingGet(ctx, id)
	return err
}

// LendingUpdate converts echo context to params.
func (w *ServerInterfaceWrapper) LendingUpdate(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.LendingUpdate(ctx, id)
	return err
}

// RepaymentGetAll converts echo context to params.
func (w *ServerInterfaceWrapper) RepaymentGetAll(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RepaymentGetAll(ctx)
	return err
}

// RepaymentCreate converts echo context to params.
func (w *ServerInterfaceWrapper) RepaymentCreate(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RepaymentCreate(ctx)
	return err
}

// UserSearch converts echo context to params.
func (w *ServerInterfaceWrapper) UserSearch(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params UserSearchParams
	// ------------- Optional query parameter "name" -------------

	err = runtime.BindQueryParameter("form", true, false, "name", ctx.QueryParams(), &params.Name)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter name: %s", err))
	}

	// ------------- Optional query parameter "email" -------------

	err = runtime.BindQueryParameter("form", true, false, "email", ctx.QueryParams(), &params.Email)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter email: %s", err))
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", ctx.QueryParams(), &params.Limit)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UserSearch(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/borrowings", wrapper.BorrowingGetAll)
	router.GET(baseURL+"/credits", wrapper.CreditsList)
	router.GET(baseURL+"/groups", wrapper.GroupGetAll)
	router.POST(baseURL+"/groups", wrapper.GroupCreate)
	router.GET(baseURL+"/groups/:id", wrapper.GroupGet)
	router.PUT(baseURL+"/groups/:id", wrapper.GroupUpdate)
	router.GET(baseURL+"/groups/:id/members", wrapper.GroupGetMembers)
	router.POST(baseURL+"/groups/:id/members", wrapper.GroupAddMember)
	router.GET(baseURL+"/health", wrapper.HealthCheck)
	router.GET(baseURL+"/lendings", wrapper.LendingGetAll)
	router.POST(baseURL+"/lendings", wrapper.LendingCreate)
	router.DELETE(baseURL+"/lendings/:id", wrapper.LendingDelete)
	router.GET(baseURL+"/lendings/:id", wrapper.LendingGet)
	router.PUT(baseURL+"/lendings/:id", wrapper.LendingUpdate)
	router.GET(baseURL+"/repayments", wrapper.RepaymentGetAll)
	router.POST(baseURL+"/repayments", wrapper.RepaymentCreate)
	router.GET(baseURL+"/users", wrapper.UserSearch)

}